+++ apps/web/src/app/photo/add/note/_components/PhotoNoteOverlay.styles.ts
