+++ apps/web/src/app/photo/add/note/_hooks/usePhotoUpload.ts
